requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
idna==2.9
praw==6.5.1
prawcore==1.0.1
python-craigslist==1.0.8
python-dotenv==0.12.0
requests==2.23.0
six==1.14.0
soupsieve==2.0
update-checker==0.16
urllib3==1.25.8
websocket-client==0.57.0
